EDGE-2529 Allow to configure additional env variables on Agent proxy for Terraform deployment (#33)
* EDGE-2529 Allow to configure additional env variables on Agent proxy for Terraform deployment

Situation

I realized that there is no way to configure different log level for Agent Proxy for ECS.

Target

We want to be able to configure that. However, ultimtely, we want to configure other env variables for Agent proxy too

Proposal.

We had several approached. The easiest one (install time) was to do the modification included.

And it allows for a customer to use this code:

```
    merge(jsondecode(module.aembit-ecs.agent_proxy_container), {
        environment = concat(module.aembit-ecs.agent_proxy_default_env_variables, [
          {"name": "AEMBIT_LOG_LEVEL", "value": "debug" }
        ]
    )}),
```
instead of

```
jsondecode(module.aembit-ecs.agent_proxy_container),
```

to specify additional env variables

* Update outputs.tf

Co-authored-by: David Walker <david.walker@quantifex.com>

---------

Co-authored-by: David Walker <david.walker@quantifex.com>